Refactor dev tools extension
